Subo las funciones de adquisicion con interrupciones
