+++ projects/EjemplosUtilesParaTPdeSapi/uart/rx_interrupt/src/rx_interrupt.h
@@ -2,9 +2,13 @@
 #define RX_INTERRUPT_H_
 #include "sapi.h"
 
+#define NoDato  0
 extern void onRx( void * );
 
-
+//static char uartBuffer[100]; //Buffer donde guardamos el string del DeltaOhm
+//static char * ptr=uartBuffer;
+//
+////static real32_t DataDeltaOhm[] = {NoDato,NoDato,NoDato,NoDato,NoDato,NoDato,NoDato};
 
 #endif
